Don't remove deleted items if copy doesn't work
part of request #9122 Improve log file for archivedeleteditem plugin
To test :
- Enable plugin
- The system works similary for Docman, FRS and Git
    * create a Docman document, a release, a Git Repository, a Svn Repository
    * delete them and launch check system event
    * the documents should be available in Admin Pending deleted documents
    * changes permissions on each item
    * change the date of purge
    * the files should not be purged on file system when the purge is launched and copy fails

+++ src/common/frs/FRSFileFactory.class.php
@@ -687,16 +687,19 @@ class FRSFileFactory {
      *
      * @return Boolean
      */
-    public function purgeFile($file, $backend) {
+    public function purgeFile($file, $backend)
+    {
         $dao = $this->_getFRSFileDao();
         if (file_exists($this->getStagingPath($file))) {
-            $this->archiveBeforePurge($file, $backend);
-            if (unlink($this->getStagingPath($file))) {
-                if (!$dao->setPurgeDate($file->getFileID(), time())) {
-                    $backend->log("File ".$this->getStagingPath($file)." not purged, Set purge date in DB fail", Backend::LOG_ERROR);
-                    return false;
+            $has_been_archived = $this->archiveBeforePurge($file, $backend);
+            if ($has_been_archived) {
+                if (unlink($this->getStagingPath($file))) {
+                    if (!$dao->setPurgeDate($file->getFileID(), time())) {
+                        $backend->log("File ".$this->getStagingPath($file)." not purged, Set purge date in DB fail", Backend::LOG_ERROR);
+                        return false;
+                    }
+                    return true;
                 }
-                return true;
             }
             $backend->log("File ".$this->getStagingPath($file)." not purged, unlink failed", Backend::LOG_ERROR);
             return false;
